local_system/requirements.txt
# ANPR Local System Dependencies
# ================================

# Core
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# AI/ML
tensorflow>=2.15.0
tf-keras>=2.15.0
ultralytics>=8.0.0
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# GUI
# tkinter is built-in to Python

# Utilities
requests>=2.31.0
huggingface-hub>=0.19.0
